Move linters from tools into tool_util
Also, update setup.py for tool_util package.

Ref: https://github.com/galaxyproject/galaxy/issues/8240

Note: `make updaterst` has not been run; waiting on
https://github.com/galaxyproject/galaxy/pull/8270
