+++ lib/galaxy/tool_util/lint.py
@@ -28,9 +28,9 @@ def lint_xml(tool_xml, level=LEVEL_ALL, fail_level=LEVEL_WARN, extra_modules=[],
 
 
 def lint_tool_source_with(lint_context, tool_source, extra_modules=[]):
-    import galaxy.tools.linters
+    import galaxy.tool_util.linters
     tool_xml = getattr(tool_source, "xml_tree", None)
-    linter_modules = submodules.import_submodules(galaxy.tools.linters, ordered=True)
+    linter_modules = submodules.import_submodules(galaxy.tool_util.linters, ordered=True)
     linter_modules.extend(extra_modules)
     for module in linter_modules:
         tool_type = tool_source.parse_tool_type() or "default"
