+++ lib/galaxy/tool_util/linters/_util.py
