+++ lib/galaxy/tool_util/linters/citations.py
