+++ lib/galaxy/tool_util/linters/command.py
