+++ lib/galaxy/tool_util/linters/general.py
