+++ lib/galaxy/tool_util/linters/inputs.py
@@ -1,6 +1,6 @@
 """This module contains a linting functions for tool inputs."""
 from galaxy.util import string_as_bool
-from ..lint_util import is_datasource
+from ._util import is_datasource
 
 
 def lint_inputs(tool_xml, lint_ctx):
