+++ lib/galaxy/tool_util/linters/outputs.py
