+++ lib/galaxy/tool_util/linters/stdio.py
